refactor: extract code to reusable component
