+++ src/index.css
@@ -4,3 +4,7 @@
   margin: 0;
   padding: 0;
 }
+
+body {
+  font-family: 'Roboto';
+}
