+++ src/pages/RegisterPage.tsx
@@ -1,31 +1,10 @@
-import { RegistrationForm } from "../components/RegistrationForm/RegistrationForm";
-import { ImageComponent } from "../components/imageComponent/ImageComponent";
-import { Box, Grid, useMediaQuery, useTheme } from "@mui/material";
+import AuthPageLayout from '../components/AuthPageLayout';
+import RegistrationForm from '../components/Forms/RegisterForm';
 
-export const RegisterPage = () => {
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down("lg"));
+export default function RegisterPage() {
   return (
-    <Box sx={{ flexGrow: 1 }}>
-      <Grid
-        container
-        spacing={2}
-        justifyContent="center"
-        alignItems="center"
-        style={{ minHeight: "100vh" }}
-      >
-        {!isMobile ? (
-          <Grid item xs={12} md={6}>
-            <ImageComponent imageURL="/images/img_cadastro.png" />
-          </Grid>
-        ) : (
-          " "
-        )}
-
-        <Grid item xs={12} md={6}>
-          <RegistrationForm />
-        </Grid>
-      </Grid>
-    </Box>
+    <AuthPageLayout imageURL="/images/img_cadastro.png">
+      <RegistrationForm />
+    </AuthPageLayout>
   );
-};
+}
